requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
jsonschema==4.21.1
openai>=2.13.0
python-dotenv>=1.0.0

